_PBKDF2_F minor optimization
get pointers to the PV* inside our SVs at the start and
pass that each iteration instead of re-extracting it each time.

for 301 decryptions and one encryption:
real    0m19.671s
to
real    0m19.658s

Perl version of the keyhash, before this, was:
real    0m32.690s

still only about 40% faster than the original.

_PBKDF2_F as this point does almost nothing, mostly
just four XORs and a method call.  that should run in a sliver
of the time of the sha_hmac256 function and this should be
four times as fast as the original, but I suspect original
profiling for the XS Digest::SHA::sha_hmac256 function may
have been undercounting.