don't use SVs for no reason
perl doesn't have a chance to re-allocate memory in that usage,
and operating on strings behind PVs is absolutely part of the API,
so I disagree that my initial version was unsafe and in anyway
in violation of promises made (Sys::Mmap is an extreme case
of this same situation).  but not creating SVs for no reason
is cleaner.  so, operate on string buffers, then only create
an SV to return.  this does still steal and use string buffers
from SVs passed as arguments, which is absolutely normal and
also not a violation of perl's API promises.